fastapi_app/requirements.txt
fastapi
uvicorn[standard]
httpx
SQLAlchemy==1.4.*
openai>=1.0.0
alembic
psycopg2-binary
pydantic
python-dotenv
aiodns
python-jose[cryptography]
